chore: post-migration cleanup — consolidate utils, remove dead code and unused deps
- Consolidate 3 duplicate asAsyncIterable() implementations into DataStream.asAsyncIterable()
- Refactor getFileAsReadStream() to return Blob, removing 8 'as any' casts in tests
- Replace 'as any' with non-null assertions in dwn-process-message tests
- Remove unused deps: randombytes (dwn-sdk-js), body-parser (dwn-server)
- Move test-only deps to devDependencies: multiformats (dwn-server), @enbox/crypto + @enbox/dids (api)
- Remove dead type exports: DwnMessageInterfaces, DwnProtocolInterfaces, DwnMessageConstructors
- Make DwnMessageWithBlob, DwnApiParams, DwnApiCreateDwnParams module-private
- Delete dead legacy example: dwn-server/examples/node/stream-data.js
- Remove stale Node.js Readable reference from stream.ts JSDoc
