+++ packages/dwn-sdk-js/package.json
@@ -86,7 +86,6 @@
     "lru-cache": "9.1.2",
     "ms": "2.1.3",
     "multiformats": "11.0.2",
-    "randombytes": "2.1.0",
     "uint8arrays": "5.1.0",
     "ulidx": "2.1.0",
     "uuid": "8.3.2",
@@ -101,7 +100,6 @@
     "@types/mocha": "^9.1.0",
     "@types/ms": "0.7.31",
     "@types/node": "^18.13.0",
-    "@types/randombytes": "2.0.0",
     "@types/secp256k1": "4.0.3",
     "@types/sinon": "^17.0.3",
     "@types/uuid": "^9.0.1",
