+++ packages/dwn-server/package.json
@@ -34,14 +34,12 @@
     "@enbox/dids": "workspace:*",
     "@enbox/dwn-sdk-js": "workspace:*",
     "@enbox/dwn-sql-store": "workspace:*",
-    "body-parser": "^1.20.2",
     "bytes": "3.1.2",
     "cors": "2.8.5",
     "express": "^4.20.0",
     "kysely": "^0.26.3",
     "loglevel": "^1.8.1",
     "loglevel-plugin-prefix": "^0.8.4",
-    "multiformats": "11.0.2",
     "mysql2": "^3.9.7",
 
     "pg": "^8.11.2",
@@ -69,6 +67,7 @@
     "crypto-browserify": "^3.12.0",
     "esbuild": "0.16.17",
     "eslint": "9.7.0",
+    "multiformats": "11.0.2",
     "eslint-plugin-mocha": "^10.4.3",
     "eslint-plugin-todo-plz": "^1.3.0",
     "http-proxy": "^1.18.1",
